Added TestUI
TestUI is a class that mimics the UI that Mrs. D created. It
exposes all of the methods from JavaArcade, and displays the game
screen.

NullGame is a simple implementation of the JavaArcade interface
that does nothing. It only exists to test TestUI and can be swaped
out easily once we have a runable prototype. Simply change the
following line:

    final static JavaArcade game = new NullGame();

in TestUI to swap out null game for another game.