Made Vector2 mutable.
The original idea behind making Vector2 imutable was so that Ball
could have a constant Vector2 to represent acceleration due to
Gravity. It never should be changed, so why sould we be able to due
so?

Creating a new Vector2 every time an operation is performed is
inefficent, and the only advantage of making a class imutable is
if we don't trust all of the classes that are using it to do the
right thing. If we don't trust the Ball class to do the right
thing with its own variables, then what piece of code could we
trust?